+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- hw-ledger, keyring, util, util-crypto 10.4.1
 - api, api-contract, types 9.13.6
 - api, api-contract, types 9.13.5
 - api, api-contract, types 9.13.4